Update favicon, change apple-touch icon, refactor About page and login route
Updated favicon and apple-touch icon in the meta tags of the frontend public index HTML. Enhanced the UI in the About page by increasing the font size and editing the style. Changed the login route to redirect to the "userdetails" page for new users. Modified the backend structure to capture and save food preferences.