Update API endpoints, remove unused files, add installation scripts
This commit includes significant changes to the structure and functioning of the application. It removes all Docker-related files reflecting a shift away from containerized deployment. API endpoints have been upgraded in both frontend and backend to use new schema and logic. Previously commented-out code in UserDetails.js is now uncommented and updated to work with the new backend. The commit also introduces two new batch scripts to automate the installation and startup process, making it easier for new users to set up the project.